requirements.txt
numpy
pandas
scikit-learn
pyyaml
nltk
lightgbm
matplotlib
seaborn
mlflow
awscli
dvc-s3
wordcloud
flask
flask_cors
pytest
boto3